OK I think this fixed it
OK so. I changed the way the paper levels module looks for the tray
levels. It looks to see if the variable is all numbers or equal to
'end'. end is what ricoh uses for a tray being empty, and integers of
25, 50, 70, and 100 for their levels. Removed SNMP test scripts, cant
get a Ricoh SNMP MIB off their site without being an employee. Removed
graphing module since I am using Google Charts to graph the data, its
easier on the server, and uses less space than creating PNG files for
each printer and graph type.