web front end changes
Forgot to upload some changes that I had made to the front end.
Changed from ordering printers by ID to name so that when you rename a
printer the ordering wont be all screwy. Still need to work on having
the daemon change the name in the printers table when it is changed in
the printers.ini file.
Added "Paper Missfeed" To the list of things it checks for to label the
printer red.
Got rid of the undefined warning when campus_id is not set in the URL.