Some more bug fixes
Changed setprinteroffline, it did not need the hostname, just its row
ID.
Created updateHostName so that when a printers name is changed in the
printers.ini config file the table will get updated also.
Also added in an exception catch for the data insert function, one of
the printers was returning a bad array sometimes and i want to see what
it is. Will remove once it is discovered.